+++ src/translate_nodes.c
@@ -6,7 +6,7 @@
 /*   By: kawish <kawish@student.codam.nl>             +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2022/04/09 14:17:39 by kawish        #+#    #+#                 */
-/*   Updated: 2022/04/19 17:58:56 by kawish        ########   odam.nl         */
+/*   Updated: 2022/05/24 10:42:23 by kgajadie      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -38,7 +38,7 @@ static unsigned int	calc_distance_rows(unsigned int rows)
 
 void	translate_nodes(t_fdf_data *fdf_data)
 {
-	unsigned int	i;
+	size_t	i;
 	unsigned int	distance_between_columns;
 	unsigned int	distance_between_rows;
